+++ changelog.md
@@ -1,5 +1,13 @@
 # Configula
 
+## 0.2.0
+
+- ADDED `read()` to read the entire configuration
+- ADDED `clone()` like read but returns a clone and not the internal configuration object
+- ADDED an error for storing objects directly
+- FIXED `isOk()` :| yes, it needed fixing
+- FIXED `write()` storing complex data types
+
 ## 0.1.2
 
 - FIXED excluding configuration files from npm deploys